+++ rust/operator-binary/src/main.rs
@@ -1,10 +1,10 @@
-use std::{ops::Deref as _, sync::Arc};
+use std::sync::Arc;
 
 use clap::Parser;
 use futures::{StreamExt, pin_mut};
 use stackable_operator::{
     YamlSchema,
-    cli::{Command, ProductOperatorRun, RollingPeriod},
+    cli::{Command, ProductOperatorRun},
     commons::authentication::AuthenticationClass,
     k8s_openapi::api::{
         apps::v1::StatefulSet,
@@ -23,9 +23,8 @@ use stackable_operator::{
     },
     logging::controller::report_controller_reconciled,
     shared::yaml::SerializeOptions,
+    telemetry::Tracing,
 };
-use stackable_telemetry::{Tracing, tracing::settings::Settings};
-use tracing::level_filters::LevelFilter;
 
 use crate::{
     crd::{
@@ -56,9 +55,6 @@ mod built_info {
 pub const APP_PORT: u16 = 8088;
 pub const OPERATOR_NAME: &str = "superset.stackable.tech";
 
-// TODO (@NickLarsenNZ): Change the variable to `CONSOLE_LOG`
-pub const ENV_VAR_CONSOLE_LOG: &str = "SUPERSET_OPERATOR_LOG";
-
 #[derive(Parser)]
 #[clap(about, author)]
 struct Opts {
@@ -82,43 +78,12 @@ async fn main() -> anyhow::Result<()> {
             telemetry_arguments,
             cluster_info_opts,
         }) => {
-            let _tracing_guard = Tracing::builder()
-                .service_name("superset-operator")
-                .with_console_output((
-                    ENV_VAR_CONSOLE_LOG,
-                    LevelFilter::INFO,
-                    !telemetry_arguments.no_console_output,
-                ))
-                // NOTE (@NickLarsenNZ): Before stackable-telemetry was used, the log directory was
-                // set via an env: `SUPERSET_OPERATOR_LOG_DIRECTORY`.
-                // See: https://github.com/stackabletech/operator-rs/blob/f035997fca85a54238c8de895389cc50b4d421e2/crates/stackable-operator/src/logging/mod.rs#L40
-                // Now it will be `ROLLING_LOGS` (or via `--rolling-logs <DIRECTORY>`).
-                .with_file_output(telemetry_arguments.rolling_logs.map(|log_directory| {
-                    let rotation_period = telemetry_arguments
-                        .rolling_logs_period
-                        .unwrap_or(RollingPeriod::Never)
-                        .deref()
-                        .clone();
-
-                    Settings::builder()
-                        .with_environment_variable(ENV_VAR_CONSOLE_LOG)
-                        .with_default_level(LevelFilter::INFO)
-                        .file_log_settings_builder(log_directory, "tracing-rs.json")
-                        .with_rotation_period(rotation_period)
-                        .build()
-                }))
-                .with_otlp_log_exporter((
-                    "OTLP_LOG",
-                    LevelFilter::DEBUG,
-                    telemetry_arguments.otlp_logs,
-                ))
-                .with_otlp_trace_exporter((
-                    "OTLP_TRACE",
-                    LevelFilter::DEBUG,
-                    telemetry_arguments.otlp_traces,
-                ))
-                .build()
-                .init()?;
+            // NOTE (@NickLarsenNZ): Before stackable-telemetry was used:
+            // - The console log level was set by `SUPERSET_OPERATOR_LOG`, and is now `CONSOLE_LOG` (when using Tracing::pre_configured).
+            // - The file log level was set by `SUPERSET_OPERATOR_LOG`, and is now set via `FILE_LOG` (when using Tracing::pre_configured).
+            // - The file log directory was set by `SUPERSET_OPERATOR_LOG_DIRECTORY`, and is now set by `ROLLING_LOGS_DIR` (or via `--rolling-logs <DIRECTORY>`).
+            let _tracing_guard =
+                Tracing::pre_configured(built_info::PKG_NAME, telemetry_arguments).init()?;
 
             tracing::info!(
                 built_info.pkg_version = built_info::PKG_VERSION,